requirements.txt
torchvision
numpy
imutils
Pillow
opencv-python
